refactor: fetch 유틸 래핑함수 추가로 인한 api 로직 병합
